only config.h is necessary to figure out #ifdef WITH_USER_THREADS tl is unused if WITH_REALTIME_THREADS is defined stack is unused
the context_switch in user_mutex_trylock was added by bdemsky to
prevent livelock.  With his stuff, it is necessary.  With a real-time
scheduler, a context switch when you're not forced to can be a bad thing
- "every thread should be guaranteed it's quanta provided it can be run" is
a mantra followed by many real-time schedulers.  Since switching in
real-time schedulers is often done when a timer expires, a context
switch is not necessary to prevent livelock.
Since I want my scheduling system to eventually be flexible enough to
accomodate any scheduling scheme (including bdemsky's) the scheduler should
be informed on a trylock in order to make good scheduling decisions.

Incidentally, when I finish this, it would be interesting to implement
bdemsky's policies in my system and see what overhead the additional
flexibility of user-defined schedulers creates.