Lots of java.io.ObjectInputStream is actually init-safe.
Runtime.insertSystemProperties is not init-safe though, since
being init-safe would involve it in an initialization cycle that
actually isn't OK (as opposed to the initialization cycles that are
OK that are specifically allowed by the JLS and break any attempts
to statically eliminate this possibly non-deterministic source of
runtime errors.)

Runtime.runFinalization is most certainly not init-safe, but we'll
pretend it's safe because we can't easily figure out what it might
depend on -> and, well, if you're finalizers can't handle a somewhat
broken (and possibly uninitialized) world, it's your own fault.
Realize, though that this means another source of runtime errors
for programs which depend on finalization.

java.net.PlainSocketImpl.available is init-safe - and maybe we might
get around to actually implementing it the right way (as opposed to
always returning a zero).

Thread.nativeInit, nativeSetPriority, and sleep are also all init-safe.