Yet another lovely option to RTJ -> thread preemption support.
Choose to either work cooperatively or preemptively with the RTJ
thread scheduler.

--with-realtime-java=threads,preempt causes preemption to occur
(sometimes in the middle of Linux kernel calls! ;-)
If our scheduler preempts a linux kernel call, when the thread switches
back, the call will either be restarted or completed (depending on
what type of call it is).  Fortunately, the newest Linux kernels are
fully reentrant and interruptable at hard 10 millisecond time boundaries.

The TimeSys kernel is interruptable at hard 50 microsecond time boundaries -
the main benefit that I can see to having an RTOS.

Be sure to also add the option -t ..._THREADS_PREEMPT
when compiling to make sure that CheckTimeSwitch() calls are not emitted.

When working preemptively, WITH_REALTIME_THREADS_PREEMPT is defined.
If it's not, then you must call CheckTimeSwitch() periodically in
order to make sure the jitter ain't too bad.

Of course, StopSwitching, StartSwitching, RestoreSwitching, setQuanta,
settimer, and context_switch all have the same behavior whether working
preemptively or non-preemptively.