New allocation instrumentation appears to work!
InstrumentedAllocationStrategy contains the only example I know about
of generating a (non-native) CALL in tree form from scratch (not just
converting a low quad PCALL intop a tree CALL).  It's actually very
neat and easy, but it's always hard to do it without any example.

New class: UnknownAllocationSiteError for the case where we want to
find the unique ID of an allocation site that was not seen by
AllocationNUmbering (previously, we were just throwing Error).  Not an
Exception as 1) it's not supposed to be caught outside the
instrumentation pacakage and 2) declaring thrown exceptions is not
that fun.