New class: AllocationStrategyFactory
Given a Frame, produces an AllocationStrategy (and we have MANY of them:
Instrumented, Preallocated, RTJ, nifty, niftystats, bdw, malloc etc.)

Remember that messy code in PreciseC.Frame for detecting the
appropriate AllocationStrategy?  I'm moving it into SAMain: SAMain
creates a Frame by passing it an appropriate
AllocationStrategyFactory, based on the command line options and the
value of "harpoon.alloc.strategy".  Hopefully, this will eliminate
many strange dependencies (why should Frame depend on command line
args and static fields scattered all over the code?)

The reason we pass a factory, and not an AllocationStrategy is that
most AllocationStrategies constructors require a frame; therefore,
they cannot be created before the call to the constructor of Frame.