More info this time: (we've just verified that cvs commit works again) Various modifications to Memory Preallocation optimization:
1. Changes to take advantage of the soon-to-be-commited modification
to BDW

2. Concentrate debug code into PreallocOpt.java

3. Add flag PreallocOpt.ONLY_SYNC_REMOVAL: run the
IncompatibilityAnalysis and remove syncs on all the preallocatable
objects, but don't actually preallocate them.  Mostly for timing
reasons: we want to measure only the difference due to memory
preallocation - the memory preallocation opt. has to involve sync
removal (otherwise, there are problems in the GC with the finalizers
for the inflated objects).