The Tree Simplification class takes a List<Rule>, not a plain list. Update PreciseGC/DynamicWBTreePass, PreciseGC/WriteBarrierConstElim, and PreciseGC/WriteBarrierTreePass to pass a correctly-parameterized list to their superclass.  Some further GJ-ization done where appropriate.
Updated gj-files and Support/gjlib.jar.

As of this commit, FLEX is fully-parameterized; that is, it ought to
compile correctly in its entirely using a proper implementation of the
JSR-14 specification for generic java.  (This doesn't mean that every
class uses parameterized types, or even that even class that *ought* to
use parameterized types does, but simply that FLEX has a consistent
type definition using in some cases "raw" types.)

In an ideal world, we would now be able to use a GJ compiler to compile
all of FLEX, without using the current two-stage compilation process.
Unfortunately, we cannot because the current JSR-14 prototype version 1.3,
although it fixes many bugs found in its predecessors, still has the
following three issues:
 1) the compiler itself throws an exception when compiling Java12.java
 2) The type-inference engine seems unable to resolve a call to
     Arrays.sort(Object[], Comparator)
    as a valid call to
      <T> void sort(T[], Comparator<T>)
    (probably according to spec, but breaks backward compatibility)
 3) A bug in compiling HArrayField/HArrayMethod, which extend
    Comparable<HMember> but not directly.

So we are sticking with our two-phase compile for the moment.  JSR-14
compiler version 1.3 -- which I encourage you to download and install from
    http://java.sun.com/people/gbracha/generics-update.html
if you are still using 1.2 or earlier -- does fix many earlier bugs, and
Sun is committed to distributing this with version 1.5 of the JDK, so I
anticipate we will not have to use our two-phase compile for long.  I may
in fact switch to a mostly-GJ compile, with special treatment only for
the 14 files which trigger the above 3 bugs.

Also, jikes and a number of other java compiler projects are moving to
support the JSR-14 GJ extensions, so I expect us to be just a little
ahead of the curve here.