MASSIVE restructuring:
1. New abstraction: CompilerState - tuple of things that constitute
the state of the compiler: main method, set of roots, code factory,
etc.  Ideally, we'll split the compiler in several stages, each taking
a CompilerState and returning a new one (to avoid errors,
CompilerState is immutable!)

2. Code generating code was moved into CodeGenerator.java, together
with its many, many options (there is a huge number of options for the
register allocator, how many of them are actually used?).  Hopefully,
other compiler stages will be moved out of SAMain.  SAMain should
consist only of option parsing and the big, top-level structure.

Many, many things remain to be done ...