1. Use appropriate directed graph terms: vertex (instead of node) and arc (instead of edge).
2. A digraph is now given by the set of roots and at least one of the
following: a forward navigator and/or a full bi-directional navigator.
The default implementation is able to infer the other one: a
bi-directional navigator is obviously a forward one too; getting a
bi-directional navigator from a forward only one is more complex as it
requires traversing the full graph to build the precedence relation.
Anyway, a bi-directional navigator is required only if it is required
(e.g., by transitivePred).

3. New graph functionality:

 3.a reverseGraph (the name should be informative enough) This occurs
 in several places in our compiler: e.g., (Meta)CallGraph
 vs. (Meta)AllCallers; now we can factors the code in one place.

 3.b transitiveSucc, transitivePred, both based on reachableVertices.
 (on long term, the separate class Reachability will be
 discontinued, and all its functionality moved in DiGraph).