Added two fields to ImageData:
scaleFactor: not used right now, but intended to store a value indicating
pixels/inch (or some similar quanitity). Should be called resolution.

trackedObjectUniqueID: integer that indicates if the object portrayed
in this ImageData is currently a tracked object. If the value is -1,
then nothing is known about the ImageData. Usually, this number only
applies when the ImageData is not full size, but cropped around a target.