Massive update to handleTracking().
If Label is told to handle tracking,
it mutates the ImageData.command tag.

Label now maintains it's own list of currently tracked objects
in addition the global ObjectTracker's list and keeps track of
how many frames that Label has found the object in.

It also keeps track of a "currently selected object".

If an object is currently selected, and if Label finds that
object again, then the object is sent on with a GO_RIGHT tag,
indicating that it is a tank and should be processed appropriately.

If the number for an object reaches a certain threshold, then the image
is sent out with  GO_LEFT tag for futher processing.