Work around broken java compilers (I think it's jikes at fault here) which don't guarantee that every MONITORENTER is matched by a MONITOREXIT. Technically, the MONITOREXIT could fail with an exception (in FLEX, a null pointer exception, but in other places also any one of a number of runtime exceptions) at which point the MONITOREXIT should *continue to be retried* until it succeeds.  Jikes allows execution to leave the method without releasing the held lock.
We work around this by assuming, for Flex's sake, that MONITOREXIT will
never fail -- not even by a NullPointerException.  This is not true
for arbitrary bytecode but *ought* to be true for bytecode generated
by legal Java programs; any NullPointerException should be thrown
before the MONITORENTER on the lock, and the compiler should guarantee
that the MONITORENTER and MONITOREXIT are using the same value.
This means that in *our* IR, at least, MONITORENTER and MONITOREXIT
are always properly matched, even if they aren't in the buggy input
bytecode.  Sigh.

You can use -Dharpoon.all.monitorexits.are.safe=true if, for some reason,
you want to disable this workaround and are confident that your input
bytecode is valid.