Bug fix: eagerly cache object fields to implement clone() method, so that later field additions aren't reflected in implementation of clone().
The problem here was that SyncTransformer was adding transaction-specific
fields to the object, which CloneImplementer was then incorrectly
trying to read and write (and then SyncTransformer was trying to
transform those reads and writes, etc, etc).  Eager generation of the
clone method fixes the problem (as does eager caching of the information
required for generation, which is the technique we actually use).