Fix the following former bug in TreeToC:     // XXX: another buglet here:  the ToTree transformation can     // create dead code by converting CJUMP(0, iffalse, iftrue) to     // JUMP(iffalse) (similarly with non-zero constants).  If this     // happens, then eventually we'll try to translate the dead     // code.  If there is a CALL in the dead code, then we'll ask     // the LiveVars about it, and the LiveVars will assert because     // the CALL is not found in its basic block set.  We should     // probably attempt to skip translation of executable (ie non-data)     // Trees which are not live. --CSA 13-mar-2003
We fix this by adding a call to Reachable, which identifies
(un)reachable code in the Tree form.  We then avoid translating
unreachable trees.  This is a little bit tricky because we don't
want to skip SEQ, DATUM, LABEL, ALIGN, etc; or sub-expressions,
like the dest and source trees of MOVE for example. I've got a
conditional which I think does the trick; a slight side-effect
is that gcc now complains about "labels at the end of compound
statements".  I don't think the warning is serious.