Removed edge/pred/succ methods from CFGrapher/CFGraphable; these have been deprecated for a while, and are replaced with edgeC/predC/succC.  The edge/pred/succ methods can't be ported to Java 1.5 because of their use of arrays of parameterized types.
Also changed the signature for edgeC/predC/succC from Collection to List,
because various bits of old code used the array returned by edge/pred/succ
as a list (i.e. depended on the consistent order of the items in the
array).  This required some cascading changes to the code.

Updated gjlib and gjlib2 after these changes.