If the Runtime is configured with the option
  --with-classpath=CLASSPATH_LIB_DIR

then we are able to find classpath glibj.zip using the directory name
${CLASSPATH_LIB_DIR} passed to the configure script.  Once we find it,
we can make sure the JAVAC compiler (either javac or jikes) uses
glibj.zip in its bootclasspath.  This way, we can use a normal
compiler, without the need for Debian + jikes-classpath.  The latest
changes in configure.in force this behaviour.

Ideally, the Flex user should be able to install Harpoon/Code,
Harppon/Runtime, classpath-install, configure Runtime by pointing it
to the correct classpath-install, and next be up and running.  The
user should not be required to commit to using Debian+jikes+classpath
as his main Java compilation/execution platform.