New code to check that the right version of classpath is on harpoon.class.path.  Currently, the user specifies which standard library implem. he wants to use (sunjdk or classpath-VER, where VER is also user-specified, with teh default being classpath-0.08).  Next, Flxe uses some heuristics to check that the right std. lib is on harpoon.class.path (currently, we grab the constant content of the field gnu.classpath.Configuration; for sunjdk, we just assume it's correct).
In the future, we may consider not requiring any command line option
about the std lib, and just look on the classpath.  The reason I'm not
inclined to do so yet is that we currently support just two versions
of classpath, and I feel that the user must be VERY AWARE of what he's
using.