Quick fix for an UNUSUALLY NASTY BUG: different implementation of HClass do not agree on the definition of equals.  For HClassProxy, equals means equality of the string descriptor.  For every single other HClass (sub)class I've seen (and there are many of them: HClass, HClassImpl, HClassCls, HClassSyn, and, last but not least!, the informatively named ImplMagic$MagicClass), so, for all of other HClass'es, equals is inherited from Object, i.e., object identity.
As a consequence, the equals relation on HClass^2 is NOT symmetric,
leading to a wide range of ugly bugs related to the use of HClass'es
in collections ...

javac should definitely contain Bill Pugh's detector of easy bugs /
suspicious code (such as defining hashCode without defining equals,
which would have caught this problem).  A couple of years ago, after a
similar debug marathon, I though about such a tool; Bill Pugh was
faster.

Currently, I've just added an equals method in ImplMagic$MagicClass.
This is a temporary fix: the best way would be to implement equals and
hashCode at the level of HClass.  Is there any reason why two
HClass'es with the same descriptor should not be equal?