Caching support added to HClass.
If you ever profiled harpoon.Analysis.Quads.CallGraphImpl in
isolation, you may have noticed how slow it is (for Jolden BH, more
than 69secs; for comparison, pointer analysis itself takes 28secs).
After investigating the issue, I've identified the problem: the
ultra-hyper-super-slow method

  HClass.getMethod(String name, String descriptor)

that does a linear search over all methods declared in a class.
Naturally, if something is slow and you call it in many places, you
try to add caching or some better search datastruct (it is trivial to
build a key tree that allows the search of a word in a huge dictionary
in time linear in the length of the searched word).  However, there is
a catch: in Flex, HClass and HMethod are mutable, so caching would be
incorrect ...

[ Darko, in case you read this, I think you once complained about the
purely functional datastructs used in some MSR compiler
infrastructure.  I think having all the important datastructs purely
functional was the best decision that MSR group ever did.  Please send
them my greetings :) ]

I think the mutability of HClass and HMethod (and the fact that the
internal representation of such classes is completely exposed -
witness the non-private, non-final, package-level "String name;" in
ClassFile/HClassCls.java) is a design mistake (especially when you
combine this mutational feature" combined with the lazy, on-demand
code factories).  However, it seems too hard to fix now and I strated
feeling really motivated to graduate soon!

As an alternative, I've added the possibility of defining "immutablity
epochs" for HClass objects: in such an epoch, all HClass'es are
guaranteed to stay unchanged, and caching makes sense.  If one wants
to do a mutation on HClass'es, the current epoch (if any) is
terminated, which renders all the HClass caches I've added irrelevant
(thanks to an epoch numbering system, these caches will get thrown
away without being used in the next immutability epoch).

The way this mechanism is supposed to be used is

 HClass.enterImmutableEpoch();

 /*
  Big computation that does not mutate HClass'es.

  E.g. call graph computation (not just the creation of the
  harpoon.Analysis.Quads.(Caching)CallGraph - that doesn't perform too
  much computation - instead, a sequence of callee computations that
  actually does something
 */

 HClass.exitImmutableEpoch();

Javadoc for these new static methods will be added soon (repeating
some of the things from this commit message).  Very soon, a commit on
the CallGraphImpl class will show a clear example of using this new
feature.