Code shuffling in src/java.lang/class.h to go around a problem encountered while compiling the Runtime with gcc (GCC) 3.4.4 20050721. In the old version, we had a prototype for
  static inline jboolean fni_class_isPrimitive

at the beginning of the file, a bunch of uses, and the real
declaration later in the file.  This works fine on most systems, but
on mine, gcc 3.4.4 20050721 gave an error of the form "inlining failed
in call to ...; function not considered for inlining".  Clearly, this
looks like a gcc bug (if the construct is illegal in C, it should tell
us so; if not, it should at most give a warning - "inline" is supposed
to be just an optimization hint).

The solution consisted in moving the declaration at the top of the
file, instead of the prototype declaration.  It seems to compile fine
now.