Restructuring: The .java files from java.lang.reflect were moved into a separate directory, reflect-thunk (these files are used to produce reflect-thunk.jar).
Motivation: Here is the way the Makefiles produced by autoconf /
automake from our various Makefile.am's work: they recurse into a
series of subdirs and execute a set of standard targets into each of
them (a target may be an executable, a native library, a .jar etc., or
a generic action like clean).  The order the different subdirs are
traversed is given in the SUBDIRS var in the parent directory-level
Makefile.am.  If subdir A contains a target that depends on a file
produced by teh Makefile from subdir B, A should be after B in
SUBDIRS.  If the dependencies between the directories have cycles,
then we have problems.

This was precisely the case here: the .java files from the old version
of the java.lang.reflect were used to generate (with javah) .h header
files that are used jni.  However, the native library built from the c
files from java.lang.reflect requires jni.

Previously, this problem was solved by having the header files in teh
CVS repository and avoiding the need to regenerate them.  However,
this is bad: why add duplicate info.

The current solution "cuts the knot" by separating between the Java
sources and the C sources from java.lang.reflect.  This way, the Java
sources can be processed first, before the C sources.