Add "fast path" versions of all the backend methods (readT, readNT, writeT, writeNT, ensureReader, ensureWriter, checkReadField, checkWriteField) and tell the compiler to inline them.  Hopefully this will enable better register allocation and optimization, so we won't be paying for unnecessary register load/saves in the common case.
We've reverted a previous commit, so that stubs is inlined now as
well.  Comparisons are still apples-to-apples.

Note that I've tagged the previous version of the Runtime with
'phd-1'; if this breaks things I can go back to the previous version
(for which I've already got a full set of numbers).